check terminated in orm
